Fixed #6: terminal lost after screen is locked
The new screenshield disables/enables all the extensions when
going to screenlock. This is a serious issue as the forked terminal is
closed.

Since the extension already rebinds nicely at startup if the terminal is
found, we just have to monkeypatch the session mode a bit to simply not
ask the terminal to quit.