Fixed #11: use gnome shell style context menu
Unfortunately, this consists in mimicking the shell style from gtk
because a MetaWindow actor can not receive events which makes opening a
popup difficult. Gnome-shell uses a virtual chrome to expose the x
windows, and doing that in the extension itself would be huge.

As font style support does not work correctly with gtk at the moment,
the menu font size is too big, but still is a better visual fit in the
shell.