Added scrollbar support
The scrollbar is at Gtk level, so it mimicks the shell look like for the
context menu.

This requires the window to be slightly transparent, and is not
compatible with the enhancement I made previously to use the terminal
transparency instead of the window one (for the font to be opaque and
more readable).

We try to do the best we can by using terminal transparency if there is
no scrollbar and falling back to window transparency if there is one.