Fixed #8: No focus given on Ctrl+Alt+Tab
... and also improve the style to match the high-res symbolic icons of
the ctrl-alt-tab popup I made recently in the shell.

To workaround the focus issue, we prevent the shell from handling our
own window and go for the manual way. This is great since the
CtrlAltTabManager.addGroup method provides a focus callback, so this
fixes the focus issue almost elegantly (this adds monkey-patching).

This requires at least gnome-shell 3.7.3 or it is a no-op.