Open the terminal faster and simplify code
There was a glitch where starting from the 2nd opening, opening the
window would play the open animation twice: first as invisible (without
us noticing) and a second time (the correct one).

We also ditch the IsOpened method of the terminal to rely on our own
tracking based on the window actor existance.

Also simplify code a lot by only relying on tweeners, even when
animation is not enabled.

This might also be a solution for #21 as we do not ask the opened state
to the terminal anymore.