Work around a gtk+ default window size regression
The gtk+ change b495ce5... introduced a regression w.r.t. to the
default window size handling.

This means the terminal has a size of 1x1 character, which is not very
useful :)

Well, we should not use the default size in the first place, but anyway.

See https://bugzilla.gnome.org/show_bug.cgi?id=696187