Fixed #38: Error upgrading extension
I should have fixed that earlier, that explains a lot of people
frustrated by the extension not working after an upgrade.

This is due to https://bugzilla.gnome.org/show_bug.cgi?id=688973 which
means we can not define twice the same GType.

This is fixed by:
- a hack for the border by suffixing by a unique global id
- proper delegation for the missing deps dialog

As discussed on IRC, gnome-shell might require a restart after upgrading
extensions in the near future to cope with that kind of issues.