Disable transparency because of Vte deprecations
The recent deprecations of 0.34.8 breaks us hard as the symbols are in
the gir file but not in the typelib, thus the symbols are not visible
from GJS, and this makes the extension crash.

Simply disable transparency until we figure out the correct way to use
transparency from now on.

Also, bump supported gnome-shell version to 3.10 for the imminent
release.