Added support for 3.9.x
I had this locally for a looong time it's nor really fair not to push
these changes even when available time is sparse.

As soon as 3.9.92 is released I'll upload for review as the review queue
is really long already, but I'm pretty sure the extension will be
rejected if we mention we support 3.9.92 but it's not even released yet.