extension: Fix terminal jumping to 2ndary monitor
This fixes the issue described in bug #45 (and to a certain degree the
one in #50).

Background: I usually have a monitor configuration with an external
monitor being the primary one and my laptop being the secondary one. The
laptop is arranged on the left, and aligned with the display bottom to
the bottom of the primary display. The external (primary) display has a
higher pixel height than the laptop (secondary) display.

The current implementation yielded that the animation started at the
primary display and then jumped to the secondary one (as described in
top of the terminal was always being cut off.

The proposed fix forces another update to the window geometry, what
resolves my problem, as the terminal is keep on the screen that contains
the top panel.