extension: Use the panelBox ::allocation signal
The panel actor allocation signal is not always sent with the various
other extensions. They almost all play with the panelBox actor instead,
like in the hide-topbar extension.

Switch to that allocation signal instead and throttles all the geometry
changes but more quickly as it seems not such a big deal comparing to
the past.

This fixes #35: Space at the top when top panel hidden