extension: Do not focus on shortcut key
I find it weird to use the shortcut key as a way to focus the terminal
as it already has a clear intent: toggling the terminal.

When you switch the focus and operate with the mouse for a short time,
you either get back using the mouse with click to focus or sloppy focus,
or you want to get rid of the whole terminal.

With this behavior, you have to think if you need to type the shortcut
key once or twice depending on the context.